Pull up dashboard object from parser
